2dcontext/imagebitmap/createImageBitmap-drawImage.html

<!DOCTYPE html>
<html>
<title>createImageBitmap + drawImage test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<script src="common.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body>
<script>
function testCanvasDisplayingPattern(canvas)
{
    var tolerance = 5; // for creating ImageBitmap from a video, the tolerance needs to be high
    _assertPixelApprox(canvas, 5,5, 255,0,0,255, "5,5", "255,0,0,255", tolerance);
    _assertPixelApprox(canvas, 15,5, 0,255,0,255, "15,5", "0,255,0,255", tolerance);
    _assertPixelApprox(canvas, 5,15, 0,0,255,255, "5,15", "0,0,255,255", tolerance);
    _assertPixelApprox(canvas, 15,15, 0,0,0,255, "15,15", "0,0,0,255", tolerance);
}

function testDrawImageBitmap(source)
{
    var canvas = document.createElement("canvas");
    canvas.width = 20;
    canvas.height = 20;
    var ctx = canvas.getContext("2d");
    ctx.clearRect(0, 0, canvas.width, canvas.height);
    return createImageBitmap(source).then(imageBitmap => {
        ctx.drawImage(imageBitmap, 0, 0);
        testCanvasDisplayingPattern(canvas);
    });
}

(function() {
    promise_test(function() {
        return makeImage().then(function(img) {
            return testDrawImageBitmap(img);
        });
    }, "createImageBitmap from a HTMLImageElement, and drawImage on the created ImageBitmap");

    promise_test(function() {
        return makeBlob().then(function(blob) {
            return testDrawImageBitmap(blob);
        });
    }, "createImageBitmap from a Blob, and drawImage on the created ImageBitmap");

    promise_test(function() {
        return makeCanvas().then(function(testCanvas) {
            return testDrawImageBitmap(testCanvas);
        });
    }, "createImageBitmap from a HTMLCanvasElement, and drawImage on the created ImageBitmap");

    promise_test(function() {
        return makeImageBitmap().then(function(bitmap) {
            return testDrawImageBitmap(bitmap);
        });
    }, "createImageBitmap from an ImageBitmap, and drawImage on the created ImageBitmap");

    promise_test(function() {
        return makeImageData().then(function(imgData) {
            return testDrawImageBitmap(imgData);
        });
    }, "createImageBitmap from an ImageData, and drawImage on the created ImageBitmap");

    promise_test(function() {
        return makeVideo().then(function(video) {
            return testDrawImageBitmap(video);
        });
    }, "createImageBitmap from a HTMLVideoElement, and drawImage on the created ImageBitmap");
})();
</script>
</body>
</html>
